requirements.txt
requests>=2.31.0
pandas>=2.2.2
SQLAlchemy>=2.0.32
PyMySQL>=1.1.1
schedule>=1.2.1
python-dotenv>=1.0.1

